remove unused dependencies, methods, and make lazy imports of conda-forge deps to improve import speed of the qligfep module
