+++ src/QligFEP/scoring.py
@@ -1,6 +1,5 @@
 """Scoring functions for perturbation network edge quality based on atom mapping similarity."""
 
-import re
 from functools import partial
 
 import numpy as np
@@ -10,7 +9,7 @@ from kartograf.atom_mapping_scorer import MappingVolumeRatioScorer
 from lomap import default_lomap_score
 from rdkit import Chem
 
-from .CLI.utils import get_avail_restraint_methods
+from .CLI.utils import parse_restraint_method
 from .logger import logger
 from .restraints.hydrogen_utils import are_hydrogens_at_end, reindex_hydrogens_to_end
 from .restraints.restraint_setter import RestraintSetter
@@ -27,53 +26,6 @@ def _ensure_hydrogens_reindexed(mol: SmallMoleculeComponent) -> SmallMoleculeCom
         rdmol = reindex_hydrogens_to_end(rdmol)
         return SmallMoleculeComponent.from_rdkit(rdmol)
     return mol
-
-
-def parse_restraint_method(restraint_method: str) -> dict:
-    """Parse a restraint method string into RestraintSetter keyword arguments.
-
-    Format: '{atom_compare}_{permissiveness}' with optional distance suffix '_{distance}'.
-    Examples: 'heavyatom_p', 'element_strict_1.2', 'kartograf'
-
-    Args:
-        restraint_method: restraint method string to parse.
-
-    Returns:
-        Dictionary with keys appropriate for RestraintSetter.set_restraints() or
-        {'kartograf_native': True} for the kartograf method.
-
-    Raises:
-        ValueError: if the method is not recognized.
-    """
-    if restraint_method == "kartograf":
-        return {"kartograf_native": True, "kartograf_max_atom_distance": 0.95}
-
-    # Check for optional atom max distance suffix
-    distance_pattern = r"_(\d+\.?\d*)$"
-    match = re.search(distance_pattern, restraint_method)
-    if match:
-        kartograf_max_atom_distance = float(match.group(1))
-        restraint_method = re.sub(distance_pattern, "", restraint_method)
-    else:
-        kartograf_max_atom_distance = 0.95
-
-    avail_methods = get_avail_restraint_methods()
-    if restraint_method not in avail_methods:
-        raise ValueError(f"Method {restraint_method} not recognized. Please use one of {avail_methods}")
-
-    atom_compare_method, permissiveness_lvl = restraint_method.split("_")
-    if permissiveness_lvl == "p":
-        params = {"strict_surround": False, "ignore_surround_atom_type": False}
-    elif permissiveness_lvl == "ls":
-        params = {"strict_surround": True, "ignore_surround_atom_type": True}
-    elif permissiveness_lvl == "strict":
-        params = {"strict_surround": True, "ignore_surround_atom_type": False}
-
-    return {
-        "atom_compare_method": atom_compare_method,
-        "kartograf_max_atom_distance": kartograf_max_atom_distance,
-        **params,
-    }
 
 
 class RestraintScorer:
